@@ -12,24 +12,30 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^4.2.4",
-    "@angular/common": "^4.2.4",
-    "@angular/compiler": "^4.2.4",
-    "@angular/core": "^4.2.4",
-    "@angular/forms": "^4.2.4",
-    "@angular/http": "^4.2.4",
-    "@angular/platform-browser": "^4.2.4",
-    "@angular/platform-browser-dynamic": "^4.2.4",
-    "@angular/router": "^4.2.4",
+    "@angular/animations": "7.1.1",
+    "@angular/common": "7.1.1",
+    "@angular/compiler": "7.1.1",
+    "@angular/core": "7.1.1",
+    "@angular/forms": "7.1.1",
+    "@angular/http": "7.1.1",
+    "@angular/platform-browser": "7.1.1",
+    "@angular/platform-browser-dynamic": "7.1.1",
+    "@angular/router": "7.1.1",
+    "bootstrap": "^4.1.3",
+    "bootstrap-notify": "^3.1.3",
+    "chartist": "^0.11.0",
     "core-js": "^2.4.1",
+    "jquery": "^3.3.1",
     "ng2-charts": "^1.6.0",
-    "rxjs": "^5.4.2",
-    "zone.js": "^0.8.14"
+    "rxjs": "^6.3.3",
+    "tslib": "^1.9.0",
+    "zone.js": "^0.8.26"
   },
   "devDependencies": {
-    "@angular/cli": "1.4.5",
-    "@angular/compiler-cli": "^4.2.4",
-    "@angular/language-service": "^4.2.4",
+    "@angular-devkit/build-angular": "~0.11.0",
+    "@angular/cli": "^7.1.1",
+    "@angular/compiler-cli": "7.1.1",
+    "@angular/language-service": "7.1.1",
     "@types/jasmine": "~2.5.53",
     "@types/jasminewd2": "~2.0.2",
     "@types/node": "~6.0.60",
@@ -42,9 +48,9 @@
     "karma-coverage-istanbul-reporter": "^1.2.1",
     "karma-jasmine": "~1.1.0",
     "karma-jasmine-html-reporter": "^0.2.2",
-    "protractor": "~5.1.2",
+    "protractor": "^5.4.1",
     "ts-node": "~3.2.0",
     "tslint": "~5.7.0",
-    "typescript": "~2.3.3"
+    "typescript": "3.1.6"
   }
 }
